refactor: move theme to UI packages
