+++ packages/ui/src/Select/index.ts
@@ -1,3 +1,4 @@
 export { default } from './Select'
 export { default as SelectLocale } from './SelectLocale'
 export { default as SelectNetwork } from './SelectNetwork'
+export { default as SelectThemes } from './SelectThemes'
