+++ packages/ui/src/utils/helpers.ts
@@ -48,3 +48,10 @@ export function shortenAccount(account: string, visibleLength = 4) {
     return account
   }
 }
+
+export const sizes = {
+  sm: '0.875rem', //14px
+  md: '1.125rem', // 18px
+  lg: '1.5rem', // 24px
+  xl: '1.75rem', // 28px
+}
